+++ blocks/browse-filters/browse-filters.css
@@ -1,23 +1,6 @@
 @import "../../scripts/browse-card/browse-cards-block.css";
 @import "../../scripts/browse-card/browse-card.css";
-
-.browse-filters-title {
-  font-weight: bold;
-  font-size: 45px;
-  line-height: 67.5px;
-  text-align: left;
-  color: var(--spectrum-gray-800);
-  margin-bottom: 8px;
-}
-
-.browse-filters-description {
-  font-weight: normal;
-  font-size: 18px;
-  line-height: 23.4px;
-  text-align: left;
-  color: var(--spectrum-gray-800);
-  margin-bottom: 32px;
-}
+@import "../browse-topics/browse-topics.css";
 
 .browse-filters-form {
   display: flex;
@@ -56,8 +39,8 @@
   line-height: 13px;
 }
 
-.filter-input > button,
-.filter-input > input {
+.filter-input>button,
+.filter-input>input {
   cursor: pointer;
   position: relative;
   height: 32px;
@@ -90,7 +73,7 @@
   height: fit-content;
 }
 
-.filter-input-search > input {
+.filter-input-search>input {
   padding-left: 35px;
   max-width: 100%;
 }
@@ -246,13 +229,13 @@ button.browse-filters-clear:disabled:hover {
 
 button.browse-tags:hover,
 button.browse-tags:focus {
-  background-color:var(--background-color);
+  background-color: var(--background-color);
   color: var(--spectrum-gray-500);
   text-decoration: underline;
   cursor: pointer;
 }
 
-.browse-filters-results > div { 
+.browse-filters-results>div {
   min-width: 0;
   min-height: 330px;
 }
@@ -332,7 +315,7 @@ button.browse-tags:focus {
   fill: var(--spectrum-gray-500);
 }
 
-.browse-filters-form .sort-container{
+.browse-filters-form .sort-container {
   align-items: center;
   color: var(--spectrum-gray-700);
   font-size: var(--spectrum-font-size-75);
@@ -343,11 +326,11 @@ button.browse-tags:focus {
   text-align: right;
 }
 
-.browse-filters-form .sort-container span{
+.browse-filters-form .sort-container span {
   vertical-align: middle;
 }
 
-.browse-filters-form .sort-container .sort-drop-btn{
+.browse-filters-form .sort-container .sort-drop-btn {
   background: transparent;
   cursor: pointer;
   color: var(--spectrum-gray-700);
@@ -362,7 +345,7 @@ button.browse-tags:focus {
   text-align: left;
 }
 
-.browse-filters-form .sort-container .sort-drop-btn:after{
+.browse-filters-form .sort-container .sort-drop-btn:after {
   background-image: url("../../icons/down-arrow.svg");
   background-position: 0 0;
   background-repeat: no-repeat;
@@ -375,7 +358,7 @@ button.browse-tags:focus {
   width: 10px;
 }
 
-.browse-filters-form .sort-container .sort-dropdown-content{
+.browse-filters-form .sort-container .sort-dropdown-content {
   background: var(--spectrum-gray-50);
   border: 1px solid var(--spectrum-gray-400);
   border-radius: 4px;
@@ -411,19 +394,19 @@ button.browse-tags:focus {
   background-size: 12px;
 }
 
-.browse-filters-form .sort-container .sort-dropdown-content a:hover{
+.browse-filters-form .sort-container .sort-dropdown-content a:hover {
   background-color: var(--spectrum-gray-100);
-} 
+}
 
-.browse-filters-form .sort-container .sort-dropdown-content.show{
+.browse-filters-form .sort-container .sort-dropdown-content.show {
   display: block;
 }
 
-.browse-filters-form.is-result .sort-container{
+.browse-filters-form.is-result .sort-container {
   display: flex;
 }
 
-.browse-filters-form:not(.is-result) .browse-filters-results{
+.browse-filters-form:not(.is-result) .browse-filters-results {
   grid-template-columns: auto;
   color: var(--spectrum-gray-700);
 }
@@ -460,7 +443,7 @@ button.browse-tags:focus {
   .browse-filters-input-container .filter-dropdown:nth-child(3) {
     grid-area: ctype;
   }
-  
+
   .browse-filters-input-container .filter-dropdown:nth-child(4) {
     grid-area: exp;
   }
