+++ blocks/browse-topics/browse-topics.js
@@ -1,12 +1,9 @@
-import { decorateIcons } from '../../scripts/lib-franklin.js';
-import { createTag, htmlToElement } from '../../scripts/scripts.js';
-
 /**
  * formattedTopicsTags returns the array of base64 encoded tags after extracting from the tags selected in dialog
  * @param {string} inputString - The topics tag. E.g. exl:topic/QXBwIEJ1aWxkZXI=
  * @returns the topic tag. E.g. QXBwIEJ1aWxkZXI=
  */
-function formattedTopicsTags(inputString) {
+export function formattedTopicsTags(inputString) {
   const splitArray = inputString.split(',');
   // eslint-disable-next-line array-callback-return, consistent-return
   const base64EncodedTagsArray = splitArray.map((item) => {
@@ -19,7 +16,7 @@ function formattedTopicsTags(inputString) {
   return base64EncodedTagsArray;
 }
 
-function handleTopicSelection(block) {
+export function handleTopicSelection(block) {
   const wrapper = block || document;
   const selectedTopics = Array.from(wrapper.querySelectorAll('.browse-topics-item-active')).reduce((acc, curr) => {
     const id = curr.dataset.topicname;
@@ -54,76 +51,4 @@ function handleTopicSelection(block) {
       window.headlessSearchEngine.dispatch(searchAction);
     }
   }
-}
-
-export default async function decorate(block) {
-  // Extracting elements from the block
-  const headingElement = block.querySelector('div:nth-child(1) > div');
-  const topics = block.querySelector('div:nth-child(2) > div').textContent.trim();
-  const allTopicsTags = topics !== '' ? formattedTopicsTags(topics) : '';
-
-  // Clearing the block's content
-  block.innerHTML = '';
-  block.classList.add('browse-topics-block');
-
-  const headerDiv = htmlToElement(`
-    <div class="browse-topics-block-header">
-      <div class="browse-topics-block-title">
-          <h2>${headingElement?.textContent.trim()}</h2>
-      </div>
-    </div>
-  `);
-  // Appending header div to the block
-  block.appendChild(headerDiv);
-  await decorateIcons(headerDiv);
-
-  const contentDiv = document.createElement('div');
-  contentDiv.classList.add('browse-topics-block-content');
-
-  allTopicsTags
-    .filter((value) => value !== undefined)
-    .forEach((topicsButtonTitle) => {
-      const topicName = atob(topicsButtonTitle);
-      const topicsButtonDiv = createTag('button', { class: 'browse-topics browse-topics-item' });
-      topicsButtonDiv.dataset.topicname = topicName;
-      // decode tags here using atob
-      topicsButtonDiv.innerHTML = topicName;
-      // click event goes here
-      contentDiv.appendChild(topicsButtonDiv);
-    });
-
-  contentDiv.addEventListener('click', (e) => {
-    if (e.target?.classList?.contains('browse-topics-item')) {
-      if (e.target.classList.contains('browse-topics-item-active')) {
-        e.target.classList.remove('browse-topics-item-active');
-      } else {
-        e.target.classList.add('browse-topics-item-active');
-      }
-      handleTopicSelection(contentDiv);
-    }
-  });
-  const decodedHash = decodeURIComponent(window.location.hash);
-  const filtersInfo = decodedHash.split('&').find((s) => s.includes('@el_features'));
-  if (filtersInfo) {
-    let selectedTopics;
-    const [, multipleFeaturesCheck] = filtersInfo.match(/@el_features==\(([^)]+)/) || [];
-    let topicsString = multipleFeaturesCheck;
-    if (!topicsString) {
-      const [, singleFeatureCheck] = filtersInfo.match(/@el_features=("[^"]*")/) || [];
-      topicsString = singleFeatureCheck;
-    }
-    if (topicsString) {
-      selectedTopics = topicsString.split(',').map((s) => s.trim().replace(/"/g, ''));
-    }
-    selectedTopics.forEach((topic) => {
-      const element = contentDiv.querySelector(`.browse-topics-item[data-topicname="${topic}"]`);
-      element.classList.add('browse-topics-item-active');
-    });
-    handleTopicSelection(contentDiv);
-  }
-  block.appendChild(contentDiv);
-
-  /* Append browse topics right above the filters section */
-  const filtersFormEl = document.querySelector('.browse-filters-form');
-  filtersFormEl.insertBefore(block, filtersFormEl.children[3]);
 }
